python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mrsf"
version = "0.3.0"
description = "CLI and library for the Markdown Review Sidecar Format (MRSF / Sidemark)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Wictor Wilén", email = "wictor@wictorwilen.se" },
]
keywords = [
    "markdown", "review", "sidecar", "mrsf", "sidemark",
    "anchoring", "comments", "cli", "yaml", "validation", "reanchor",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1",
    "jsonschema>=4.20",
    "ruamel.yaml>=0.18",
    "rapidfuzz>=3.6",
    "rich>=13.7",
    "watchdog>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.9",
]

[project.scripts]
mrsf = "mrsf.cli.main:cli"

[project.urls]
Homepage = "https://github.com/wictorwilen/MRSF"
Repository = "https://github.com/wictorwilen/MRSF"
Issues = "https://github.com/wictorwilen/MRSF/issues"
Documentation = "https://sidemark.org"

[tool.hatch.build.targets.wheel]
packages = ["src/mrsf"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "mrsf.cli.*"
disallow_untyped_decorators = false
disable_error_code = ["has-type"]

[tool.pytest.ini_options]
testpaths = ["tests"]
